+++ src/main/scala/clustered/ClusterDescriptor.scala
@@ -4,7 +4,7 @@ import `object`.Object
 import math.Vec2
 
 case class ClusterDescriptor(id: String,
-                             mass: BigDecimal,
+                             mass: Double,
                              _position: Vec2) extends Object {
   override var position: Vec2 = _position
 }
